requirements.txt
apparmor==3.0.1
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.3.1
asn1crypto==1.4.0
Brotli==1.0.9
btrfsutil==5.10.1
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.4.6
cupshelpers==1.0
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
distlib==0.3.1
distro==1.5.0
Django==3.1.7
djangorestframework==3.12.2
docopt==0.6.2
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
html5lib==1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
keyutils==0.6
LibAppArmor==3.0.1
louis==3.16.0
MarkupSafe==1.1.1
meson==0.57.1
msgpack==1.0.2
netsnmp-python==1.0a1
npyscreen==4.10.5
numpy==1.20.1
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.19.2
pandas==1.2.3
pep517==0.9.1
Pillow==8.1.0
plotly==4.14.3
ply==3.11
progress==1.5
pycairo==1.20.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
PyGObject==3.38.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.3
PyQt5-sip==12.8.1
pysmbc==1.0.23
python-dateutil==2.8.1
python-libtorrent==1.2.11
pytz==2021.1
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
reportlab==3.5.60
requests==2.25.1
resolvelib==0.5.4
retrying==1.3.3
rgw==2.0.0
six==1.15.0
sqlparse==0.4.1
team==1.0
toml==0.10.2
udiskie==2.3.2
urllib3==1.26.3
webencodings==0.5.1
Werkzeug==1.0.1
